requirements.txt
python-telegram-bot==22.5
python-telegram-bot[job-queue]==22.5
aiofiles==23.1.0
Pillow==11.3.0
requests
pyrofork
TgCrypto-pyrofork